Updating tables and UX

+++ front/src/app/home/home.component.ts
@@ -4,10 +4,13 @@ import { first } from 'rxjs/operators';
 import { Student } from '@/_models';
 import { StudentService } from '@/_services';
 import { Router } from '@angular/router';
+import { Subject } from 'rxjs';
 
 @Component({ templateUrl: 'home.component.html' })
 export class HomeComponent implements OnInit {
     students: Student[] = [];
+    dtTrigger: Subject<any> = new Subject<any>();
+    dtOptions: DataTables.Settings = {};
 
     constructor(
         private studentService: StudentService,
@@ -15,6 +18,21 @@ export class HomeComponent implements OnInit {
         ) {}
 
     ngOnInit() {
+        this.dtOptions = {
+            pagingType: 'full_numbers',
+            pageLength: 2,
+            language: {
+                lengthMenu: "Linhas por página: _MENU_",
+                info: "_START_ - _END_ de _TOTAL_",
+                paginate: {
+                    first:      "<<",
+                    last:       ">>",
+                    next:       ">",
+                    previous:   "<"
+                }
+            }
+        };
+
         this.loadAllStudents();
     }
 
@@ -22,10 +40,20 @@ export class HomeComponent implements OnInit {
         this.router.navigate(['/register'], { queryParams: { student: JSON.stringify(student) } });
     }
 
+    studentDetail(student) {
+        this.router.navigate(['/detail'], { queryParams: { student: JSON.stringify(student) } });
+    }
+
     deleteStudent(id: number) {
-        this.studentService.delete(id).pipe(first()).subscribe(() => {
-            this.loadAllStudents()
-        });
+        var r = confirm("Deseja realmente deletar o Aluno?");
+        if (r == true) {
+            this.studentService.delete(id).pipe(first()).subscribe(() => {
+                this.loadAllStudents()
+            });
+        } else {
+            console.log("Aluno não deletado");
+        }
+        
     }
 
     searchStudentsByParam(searchValue: string) {
@@ -45,6 +73,7 @@ export class HomeComponent implements OnInit {
     private loadAllStudents() {
         this.studentService.getAll().pipe(first()).subscribe(students => {
             this.students = students;
+            this.dtTrigger.next();
         });
     }
 }
